Clique cutset unittest
Moved the unit test to its own file.
