+++ inducer/clique_cutset.py
@@ -11,10 +11,6 @@ Version: 2015-10-21
 """
 from networkx import is_connected, find_cliques
 from itertools import combinations
-from inducer.helper import make_clique, make_diamond, make_cycle, make_kite,\
-    make_bridge
-import unittest
-import networkx as nx
 
 
 def clique_cutset(G):
@@ -52,34 +48,3 @@ def subset(clique):
     for i in range(1, len(clique) + 1):
         for x in combinations(clique, i):
             yield x
-
-
-class Test(unittest.TestCase):
-
-    def testCliqueCutset(self):
-        # no cutset
-        result = clique_cutset(make_cycle(5))
-        self.assertEqual(result, None)
-        # cutset is subclique of the maximal clique
-        result = clique_cutset(make_diamond())
-        self.assertEqual(result.nodes(), make_clique(2).nodes())
-        self.assertEqual(result.edges(), make_clique(2).edges())
-
-        # just a normal cutset
-        g = nx.Graph()
-        g.add_node(2)
-        result = clique_cutset(make_kite())
-        self.assertEqual(result.nodes(), g.nodes())
-        # whole graph is a clique
-        result = clique_cutset(make_clique(4))
-        self.assertEqual(result.nodes(), make_clique(4).nodes())
-        self.assertEqual(result.edges(), make_clique(4).edges())
-        # a random graph
-        result = clique_cutset(make_bridge())
-        self.assertEqual(result.nodes(), make_clique(2).nodes())
-        self.assertEqual(result.edges(), make_clique(2).edges())
-
-
-if __name__ == "__main__":
-    # import sys;sys.argv = ['', 'Test.testName']
-    unittest.main()
